requirements.txt
dnspython==2.7.0
tldextract==5.1.3
